requirements.txt
asgiref==3.5.2
distlib==0.3.5
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
filelock==3.8.0
pbr==5.10.0
platformdirs==2.5.2
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
stevedore==4.0.0
tzdata==2022.1
whitenoise==6.2.0